fix: 2번째 PR 피드백 반영
