+++ android/app/src/main/java/com/example/issue_tracker/repository/RepositoryModule.kt
@@ -15,4 +15,10 @@ abstract class RepositoryModule {
     abstract fun bindLabelRepository(
         labelRepositoryImpl: LabelRepositoryImpl
     ): LabelRepository
+
+    @Singleton
+    @Binds
+    abstract fun bindIssueRepository(
+        issueRepositoryImpl: IssueRepositoryImpl
+    ): IssueRepository
 }
