+++ android/app/src/main/java/com/example/issue_tracker/ui/HomeActivity.kt
@@ -10,26 +10,26 @@ import androidx.databinding.DataBindingUtil
 import androidx.navigation.findNavController
 import androidx.navigation.ui.setupWithNavController
 import com.example.issue_tracker.R
-import com.example.issue_tracker.databinding.ActivityIssueBinding
+import com.example.issue_tracker.databinding.ActivityHomeBinding
 import com.example.issue_tracker.ui.issue.IssueFragment
 import com.example.issue_tracker.ui.issue.IssueViewModel
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class IssueActivity : AppCompatActivity() {
+class HomeActivity : AppCompatActivity() {
 
-    private lateinit var binding: ActivityIssueBinding
+    private lateinit var binding: ActivityHomeBinding
     private val viewModel: IssueViewModel by viewModels()
     private val issueFragment = IssueFragment()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = DataBindingUtil.setContentView(this, R.layout.activity_issue)
+        binding = DataBindingUtil.setContentView(this, R.layout.activity_home)
     }
 
     override fun onStart() {
         super.onStart()
-        binding.issueBottomNavigation.setupWithNavController(findNavController(R.id.nav_issue_fragment))
+        binding.issueBottomNavigation.setupWithNavController(findNavController(R.id.nav_home_fragment))
     }
 
     // 화면 아무 곳이나 클릭 하면 키보드 숨김
