+++ android/app/src/main/java/com/example/issue_tracker/ui/common/SwipeHelperCallback.kt
@@ -1,6 +1,7 @@
 package com.example.issue_tracker.ui.common
 
 import android.graphics.Canvas
+import android.util.Log
 import android.view.View
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.ItemTouchHelper.*
