+++ android/app/src/main/java/com/example/issue_tracker/ui/issue/IssueViewModel.kt
@@ -4,6 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.issue_tracker.model.Issue
 import com.example.issue_tracker.repository.IssueRepository
+import com.example.issue_tracker.repository.IssueRepositoryImpl
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
