+++ android/app/src/main/res/layout/activity_home.xml
@@ -10,10 +10,10 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.IssueActivity">
+        tools:context=".ui.HomeActivity">
 
         <androidx.fragment.app.FragmentContainerView
-            android:id="@+id/nav_issue_fragment"
+            android:id="@+id/nav_home_fragment"
             android:name="androidx.navigation.fragment.NavHostFragment"
             android:layout_width="match_parent"
             android:layout_height="0dp"
